+++ src/app/admin/posts/_components/PostForm.tsx
@@ -1,23 +1,23 @@
 "use client";
 
-import { useEffect, ChangeEvent } from "react";
-import { useForm, Controller, useWatch } from "react-hook-form";
+import { useEffect } from "react";
+import { useForm, Controller } from "react-hook-form";
 import Input from "@/app/contact/form/Input";
 import Label from "@/app/contact/form/Label";
 import { Post } from "@/app/_types/post";
-import { supabase } from "@/utils/supabase";
-import { v4 as uuidv4 } from "uuid";
-import Image from "next/image";
 import { useAdminCategories } from "../../categories/_hooks/useAdminCategories";
+import PostImageUploader from "./PostImageUploader"; // 画像アップローダーをインポート
+
+type PostFormData = {
+  title: string;
+  content: string;
+  thumbnailImageKey: string;
+  categories: number[];
+};
 
 type PostFormProps = {
   initialData?: Post;
-  onSubmit: (postData: {
-    title: string;
-    content: string;
-    thumbnailImageKey: string;
-    categories: number[];
-  }) => void;
+  onSubmit: (postData: PostFormData) => void;
   onDelete?: () => void;
   buttonText: string;
 };
@@ -36,7 +36,7 @@ const PostForm: React.FC<PostFormProps> = ({
     handleSubmit,
     setValue,
     formState: { errors, isSubmitting },
-  } = useForm({
+  } = useForm<PostFormData>({
     defaultValues: {
       title: initialData?.title || "",
       content: initialData?.content || "",
@@ -44,15 +44,6 @@ const PostForm: React.FC<PostFormProps> = ({
       categories: initialData?.categories?.map((category) => category.id) || [],
     },
   });
-
-  // useWatch でリアルタイム監視
-  const thumbnailImageKey = useWatch({ control, name: "thumbnailImageKey" });
-
-  // サムネイル画像のURLを取得
-  const thumbnailImageUrl = thumbnailImageKey
-    ? supabase.storage.from("post_thumbnail").getPublicUrl(thumbnailImageKey)
-        .data.publicUrl
-    : null;
 
   // 初期データがある場合にフォームにセットする
   useEffect(() => {
@@ -66,25 +57,6 @@ const PostForm: React.FC<PostFormProps> = ({
       );
     }
   }, [initialData, setValue]);
-
-  // 画像ファイルが選択された場合の処理
-  const handleImageChange = async (event: ChangeEvent<HTMLInputElement>) => {
-    if (!event.target.files || event.target.files.length === 0) return;
-
-    const file = event.target.files[0];
-    const filePath = `private/${uuidv4()}`;
-
-    const { error } = await supabase.storage
-      .from("post_thumbnail")
-      .upload(filePath, file, { cacheControl: "3600", upsert: false });
-
-    if (error) {
-      alert(error.message);
-      return;
-    }
-
-    setValue("thumbnailImageKey", filePath); // useForm に値をセット
-  };
 
   return (
     <div className="max-w-[800px] mx-auto py-10">
@@ -119,32 +91,8 @@ const PostForm: React.FC<PostFormProps> = ({
           )}
         </div>
 
-        {/* 現在のサムネイル画像があれば表示 */}
-        {thumbnailImageUrl && (
-          <div className="flex flex-col mb-6">
-            <Label htmlFor="thumbnailImageKey">現在のサムネイル画像</Label>
-            <div className="relative w-24 h-24 mb-4">
-              <Image
-                src={thumbnailImageUrl}
-                alt="サムネイル画像"
-                layout="fill"
-                className="object-cover rounded-lg"
-              />
-            </div>
-          </div>
-        )}
-
-        {/* 新しいサムネイル画像アップロード */}
-        <div className="flex flex-col mb-6">
-          <Label htmlFor="thumbnailImageKey">サムネイル画像</Label>
-          <input
-            id="thumbnailImageKey"
-            type="file"
-            accept="image/*"
-            onChange={handleImageChange}
-            className="p-2 border rounded-lg"
-          />
-        </div>
+        {/* 画像アップローダー */}
+        <PostImageUploader control={control} setValue={setValue} />
 
         {/* カテゴリ選択 */}
         <div className="flex flex-col mb-6">
